Refactor for major time savings running refs.py.
Before: 2127 seconds.
After: 817 seconds.

The gff2 and gff3 annotation files are the source of all the data processed in refs.py. The gff2 file is ~3GBs in size and gff3 file is ~4GBs.

Before this change, we were reading the gff2 file 9 independent times and the gff3 file 3 independent times. Since I/O operations are expensive (in general), minimizing the amount of time reading files from disk is a good way to improve performance. So if we cut down those 9 reads of the file to 1, we'll get really nice time savings.

That's what this change does. We now read each file only once. We just do the 9 things all at once (when we read each record) instead of in 9 separate loops.